+++ library/SolverUtils/Diffusion/Diffusion.h
@@ -350,7 +350,7 @@ namespace Nektar
             /// Get the average and jump value of conservative variables on trace
             SOLVER_UTILS_EXPORT void ConsVarAveJump(
                 const std::size_t                             nConvectiveFields,
-                const int                                           npnts,
+                const size_t                                        npnts,
                 const Array<OneD, const Array<OneD, NekDouble> >    &vFwd,
                 const Array<OneD, const Array<OneD, NekDouble> >    &vBwd,
                       Array<OneD,       Array<OneD, NekDouble> >    &aver,
@@ -403,7 +403,7 @@ namespace Nektar
         
             SOLVER_UTILS_EXPORT virtual void v_ConsVarAveJump(
                 const std::size_t                             nConvectiveFields,
-                const int                                           npnts,
+                const size_t                                        npnts,
                 const Array<OneD, const Array<OneD, NekDouble> >    &vFwd,
                 const Array<OneD, const Array<OneD, NekDouble> >    &vBwd,
                       Array<OneD,       Array<OneD, NekDouble> >    &aver,
